Moves headings to their own file - issue #670
Question: Should form-specific heading things be moved to a
different directory? A file in the 'forms' folder? I know we're
not using `ContentH1` anywhere other than forms right now, but
I hope to use it in the About page and maybe in our style guide,
other documentation, etc.
